+++ src/root/controller/Debugger.java
@@ -39,8 +39,8 @@ public final class Debugger {
     }
 
     private void setSnake() {
-        Trail data = recorder.getData((byte) 0);
-        Snake snake = board.getSnake((byte) 0);
+        Trail data = recorder.getData(0);
+        Snake snake = board.getSnake(0);
 
         short length = data.getLength(nextStep);
         int tailIndex = currentStep-length;
@@ -78,7 +78,7 @@ public final class Debugger {
         snake.setTail(snake.getBody().get(length-1));
 
         if (object != status.STILL)
-            recorder.printValue((byte) 0, nextStep, object);
+            recorder.printValue(0, nextStep, object);
 //        currentStep = nextStep;
         }
 
@@ -88,13 +88,13 @@ public final class Debugger {
 //        recorder.removeData(
 //                board.getSnake((byte) 0).getSerialNum(),
 //                stepNow);
-        recorder.removeData((byte) 0, stepNow);
+        recorder.removeData(0, stepNow);
         backSteps = 0;
         object = status.STILL;
         tuned = false;
     }
     private len snakeExtended(int currentStep, int nextStep) {
-        Trail data = recorder.getData((byte) 0);
+        Trail data = recorder.getData(0);
         int currentLength = data.getLength(currentStep);
         int nextLength = data.getLength(nextStep);
         if (currentLength < nextLength)
@@ -105,7 +105,7 @@ public final class Debugger {
     }
 
     private void setFood() {
-        Trail value = recorder.getData((byte) 0);
+        Trail value = recorder.getData(0);
         board.getFood().setPoint(value.getFood(currentStep));
 //        board.getFood() = new Food(value.getFood(currentStep));
     }
